requirements.txt
2captcha-python==1.1.3
allure-pytest==2.12.0        
allure-python-commons==2.12.0
async-generator==1.10        
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1    
colorama==0.4.6
comtypes==1.1.14
exceptiongroup==1.0.2        
h11==0.14.0
html-testRunner==1.2.1       
idna==3.4
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyparsing==3.0.9
pypiwin32==223
PySocks==1.7.1
pytest==7.2.0
pytest-html==3.2.0
pytest-metadata==2.0.4       
python-dotenv==0.21.0        
pywin32==305
requests==2.28.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0      
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2        
urllib3==1.26.12
webdriver-manager==3.8.4     
wsproto==1.2.0